flask/requirements.txt
bcrypt==3.1.4
cffi==1.11.5
click==6.7
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
Pillow==5.2.0
pycparser==2.18
six==1.11.0
SQLAlchemy==1.2.11
Werkzeug==0.14.1
WTForms==2.2.1
